WebhookHandler.php: Use '#'.$issue['iid'] as the reference.
I.e., the "internal" = per-project ID, instead of $issue['id'], i.e.,
the global ID, matching what the import script
(https://github.com/kkofler/kanboard-import-gitlab) does.

This fixes compatibility with the import script, makes the references
user-readable, and is robust when a project gets migrated to a different
GitLab instance. The drawback is that the iid is (by design) not
globally unique, so issues moved from one project to another can confuse
the webhook.